change the size of board size from byte to int the si

+++ ReverseTicTacToeGame/ValiditionUI.cs
@@ -4,8 +4,10 @@ namespace ReverseTicTacToeGame
 {
     class ValiditionUi
     {
-        public static byte GetValidBoardSize()
+        public static int GetValidBoardSize()
         {
+            //get number between 3-9
+            //return number between 3-9 
             throw new NotImplementedException();
         }
